+++ application/language/bulgarian/translations_lang.php
@@ -510,4 +510,6 @@ $lang['no'] = 'No';
 $lang['latest_blog_posts'] = 'Latest Blog Posts';
 $lang['loading'] = 'Loading';
 $lang['blog_posts_load_error'] = 'Unable to load blog posts. Please visit the official website for updates.';
+$lang['would_you_like_to_proceed'] = 'Would you like to proceed anyway?';
+$lang['proceed'] = 'Proceed';
 // End