+++ assets/js/utils/calendar_default_view.js
@@ -130,7 +130,7 @@ App.Utils.CalendarDefaultView = (function () {
                         );
                     },
                 );
-            } else if (!lastFocusedEventData.extendedProps.data.is_unavailability) {
+            } else if (!Number(lastFocusedEventData.extendedProps.data.is_unavailability)) {
                 const appointment = lastFocusedEventData.extendedProps.data;
 
                 App.Components.AppointmentsModal.resetModal();
@@ -249,7 +249,7 @@ App.Utils.CalendarDefaultView = (function () {
                 const date = moment(lastFocusedEventData.start).format('YYYY-MM-DD');
 
                 App.Http.Calendar.deleteWorkingPlanException(date, providerId, successCallback);
-            } else if (!lastFocusedEventData.extendedProps.data.is_unavailability) {
+            } else if (!Number(lastFocusedEventData.extendedProps.data.is_unavailability)) {
                 const buttons = [
                     {
                         text: lang('cancel'),
@@ -819,7 +819,7 @@ App.Utils.CalendarDefaultView = (function () {
             $notification.hide('bind');
         }
 
-        if (!info.event.extendedProps.data.is_unavailability) {
+        if (!Number(info.event.extendedProps.data.is_unavailability)) {
             // Prepare appointment data.
             info.event.extendedProps.data.end_datetime = moment(info.event.extendedProps.data.end_datetime)
                 .add({days: info.endDelta.days, milliseconds: info.endDelta.milliseconds})
@@ -868,14 +868,28 @@ App.Utils.CalendarDefaultView = (function () {
                     text: lang('no'),
                     click: (event, messageModal) => {
                         messageModal.hide();
-                        App.Http.Calendar.saveAppointment(appointment, null, () => successCallback(false), null, false);
+                        App.Http.Calendar.saveAppointmentWithConflictHandling(
+                            appointment,
+                            null,
+                            () => successCallback(false),
+                            null,
+                            false,
+                            () => info.revert(),
+                        );
                     },
                 },
                 {
                     text: lang('yes'),
                     click: (event, messageModal) => {
                         messageModal.hide();
-                        App.Http.Calendar.saveAppointment(appointment, null, () => successCallback(true), null, true);
+                        App.Http.Calendar.saveAppointmentWithConflictHandling(
+                            appointment,
+                            null,
+                            () => successCallback(true),
+                            null,
+                            true,
+                            () => info.revert(),
+                        );
                     },
                 },
             ]);
@@ -970,7 +984,7 @@ App.Utils.CalendarDefaultView = (function () {
 
         let successCallback;
 
-        if (!info.event.extendedProps.data.is_unavailability) {
+        if (!Number(info.event.extendedProps.data.is_unavailability)) {
             // Prepare appointment data.
             const appointment = {...info.event.extendedProps.data};
 
@@ -1031,14 +1045,28 @@ App.Utils.CalendarDefaultView = (function () {
                     text: lang('no'),
                     click: (event, messageModal) => {
                         messageModal.hide();
-                        App.Http.Calendar.saveAppointment(appointment, null, () => successCallback(false), null, false);
+                        App.Http.Calendar.saveAppointmentWithConflictHandling(
+                            appointment,
+                            null,
+                            () => successCallback(false),
+                            null,
+                            false,
+                            () => info.revert(),
+                        );
                     },
                 },
                 {
                     text: lang('yes'),
                     click: (event, messageModal) => {
                         messageModal.hide();
-                        App.Http.Calendar.saveAppointment(appointment, null, () => successCallback(true), null, true);
+                        App.Http.Calendar.saveAppointmentWithConflictHandling(
+                            appointment,
+                            null,
+                            () => successCallback(true),
+                            null,
+                            true,
+                            () => info.revert(),
+                        );
                     },
                 },
             ]);